Check for ValueError in write only check in _unwrap_text()
In the _unwrap_text() function it checks by trying to read or write to
the stream to see if it's wrapped or not and return the buffer. The
check there first tries to read and if an exception of the proper type
is raised it will assume it was opened in write mode and proceed to
check for a buffer by writing to the stram. However this check is
missing an exception to check for. If a FileIO object is passed in on
Python 2.7 (which can be the case sometimes for sys.stdout, although I
haven't figured out under which circumstances) then a ValueError will be
raised. [1] To enable using _unwrap_text() when a FileIO object is used
this commit adds a ValueError to the list of exception types to handle
this case.

[1] https://github.com/python/cpython/blob/2.7/Modules/_io/fileio.c#L431