Switch to setuptools entrypoint console scripts for filters
This commit switches from using scripts to setuptools entrypoint console
scripts for the filter scripts. This should fix the issues for windows
users trying to execute the scripts. To accomplish this the filter
scripts are moved into the subunit package namespace and a callable to
run the script is added to each module.

Closes-Bug: #1322888