Add support for new python versions and update pkg metadata
This commit updates the supported python versions for subunit to include
support for newer versions of the language 3.7 and 3.8 which have been
released since the last python-subunit release. At the same time it
takes the opportunity to update the package metadata to add more details
including the new python versions.