Work around 'short read' race
We parse packets inside a loop. When there's a large amount of captured
data, we see error messages like the following:

  Captured Packet data:
  ~~~~~~~~~~~~~~~~~~~~~
      b'\xb3+`Pr^'

  Captured Parser Error:
  ~~~~~~~~~~~~~~~~~~~~~~
      b'Short read - got 108 bytes, wanted 4204 bytes'

Interestingly, if you scan through the logs, you'll see the following
combination popping up (varies between test run):

  Short read - got 108 bytes, wanted 4204 bytes
  Short read - got 4090 bytes, wanted 4204 bytes

These come in the wrong order because things are running in parallel,
but it's obvious that we're not waiting long enough to parse things and
are racing as a result. The default type of 'sys.stdout' in Python 3 is
'io.TextIOBase'. The 'read' method of this returns "at most size
characters from the stream as a single str" [1]. The solution seems
obvious - if we don't get enough packets, try again before we actually
bail out.

[1] https://docs.python.org/3/library/io.html#io.TextIOBase.read

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Partial-bug: #1813147