Pivot CI to github actions
With the tight travis quotas for open source projects now it is not a
viable provider for running CI on this project anymore. While this
project is relatively low activity the quota seems to be preventing
jobs from reliably triggering. This commit switches CI to use github
actions which doesn't have this limitation.