Drop support for Python 2.7, 3.5
Both of these are EOL now. There's no reason to continue supporting
them. Python 3.6 is close to EOL but we can give it another release
before we formally drop that.

We also drop support for installing without setuptools, since this is
expected to be effectively always present nowadays. We include a
pyproject.toml file just in case that ever changes though.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>