Remove perl support from the main subunit repo
This has now moved to https://github.com/jelmer/subunit-perl

This simplifies the build system logic and makes it unnecessary to
integrate auto{conf,make} and the Perl buildsystem. The perl support
is also not in the samestate as the rest of the code - it doesn't have a
testsuite, and only supports subunit v1.